create auth.ts
